requirements.txt
fastapi
uvicorn
uvloop==0.14.0
httptools
asgiref==3.3.4
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
Werkzeug==1.0.1
Pillow==8.2.0
numpy==1.19.5
